// frontend/src/app/core/api/model/training-session-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TrainingSessionResponse { 
    id?: number;
    promotionName?: string;
    academicYearId?: number;
    academicYearLabel?: string;
    enrolledCount?: number;
    expectedStartDate?: string;
    estimatedEndDate?: string;
    status?: TrainingSessionResponse.StatusEnum;
    levelId?: number;
    levelCode?: string;
    specialtyId?: number;
    specialtyCode?: string;
    registrationOpenDate?: string;
    registrationDeadline?: string;
    registrationsOpen?: boolean;
}
export namespace TrainingSessionResponse {
    export const StatusEnum = {
        InProgress: 'IN_PROGRESS',
        Finished: 'FINISHED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}